Organise AJAX code with new module
