+++ public/src/requests.js
@@ -1,0 +1,19 @@
+export async function getYears() {
+  // TODO
+}
+
+export function getYear(year) {
+  const xhttp = new XMLHttpRequest();
+  xhttp.onreadystatechange = function () {
+    // Successful request
+    if (this.readyState === 4 && this.status === 200) {
+      const res = JSON.parse(this.responseText);
+      console.log(res);
+    }
+  };
+
+  xhttp.open('POST', '/year', true);
+  xhttp.setRequestHeader("Content-Type", "application/json");
+
+  xhttp.send(JSON.stringify({ year }));
+}
